python-scrapers/requirements.txt
pytrends>=4.9.2
psycopg2-binary>=2.9.9

